Advanced_FL/requirements.txt
torch>=2.2
torchvision>=0.17
numpy>=1.26
pandas>=2.0
scikit-learn>=1.4
scipy>=1.11
tqdm>=4.66
pyyaml>=6.0
opacus>=1.4
mediapipe>=0.10
opencv-python>=4.9
matplotlib>=3.8
jupytext>=1.16